Reorganizing lots of files in preparation for scores

+++ lib/src/common/doc_coverage.dart
@@ -106,7 +106,7 @@ class DocCoverage {
           }
 
           String comment = resolveCommentText(thing['comment']);
-          int score = 1.0;
+          double score = 1.0;
           if (comment.isEmpty)
             return 0.0;
